mm-video: Add support for ouput buffers color conversion.
Video decoder hardware generates YUV in tile format.
Client of openmax video component can request for a different
color format in allocate buffer model.

+++ mm-video/vidc/vdec/inc/mp4_utils.h
@@ -30,11 +30,11 @@ ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "OMX_Core.h"
 #include "OMX_QCOMExtns.h"
 typedef signed long long int64;
-typedef unsigned long int uint32;   /* Unsigned 32 bit value */
+typedef unsigned int uint32;   /* Unsigned 32 bit value */
 typedef unsigned short uint16;   /* Unsigned 16 bit value */
 typedef unsigned char uint8;   /* Unsigned 8  bit value */
 
-typedef signed long int int32;   /* Signed 32 bit value */
+typedef int int32;   /* Signed 32 bit value */
 typedef signed short int16;   /* Signed 16 bit value */
 typedef signed char int8;   /* Signed 8  bit value */
